First commit
Created edit folder
Created initialize.php
Creating $URLDepth logic
