+++ edit/characters.php
@@ -1,30 +1,50 @@
 <!doctype php>
 <?php
-	include_once "server/conexao.php";
-	include_once "server/functions.php";
+	$URLDepth = '../';
+	require_once "../server/initialize.php";
 	if(!empty($_POST)){
 		$_POST = array_map("strip_tags", $_POST);
 		$_GET = array_map("strip_tags", $_GET);
+		//print_r($_POST);
 		verify_password($conn);
 		
 		if($_POST['action'] == 'Update'){
 		
-			$query = "UPDATE `game_entry` SET `title`= ?, `order` = ? WHERE `entryid` = ?";
+			$query = "UPDATE `character` SET `Name`=? WHERE `idcharacter` = ?";
 			$result = $conn -> prepare($query);
-			$result -> bind_param("sii", $_POST['entry'], $_POST['order'], $_POST['entryid']);
+			//print_r($_POST);
+			$result -> bind_param("si", $_POST['character'], $_POST['idcharacter']);
 			$result -> execute();
 		}else if($_POST['action'] == 'Delete'){
-			
-			$query = "DELETE FROM `game_entry` WHERE `entryid` = ?";
+			$query = "DELETE `resources` FROM `character` 
+JOIN `combo` ON `combo`.`character_idcharacter` = `character`.`idcharacter`
+JOIN `resources` ON `resources`.`combo_idcombo` = `combo`.`idcombo`
+WHERE `character`.`idcharacter` = ?";
 			$result = $conn -> prepare($query);
-			$result -> bind_param("i", $_POST['entryid']);
+			//print_r($_POST);
+			$result -> bind_param("i", $_POST['idcharacter']);
+			$result -> execute();
+			
+			$query = "DELETE `combo` FROM `character` 
+JOIN `combo` ON `combo`.`character_idcharacter` = `character`.`idcharacter`
+WHERE `character`.`idcharacter` = ?";
+			$result = $conn -> prepare($query);
+			//print_r($_POST);
+			$result -> bind_param("i", $_POST['idcharacter']);
+			$result -> execute();
+			
+			$query = "DELETE FROM `character` WHERE `idcharacter` = ?";
+			$result = $conn -> prepare($query);
+			//print_r($_POST);
+			$result -> bind_param("i", $_POST['idcharacter']);
 			$result -> execute();
 			
 			
 		}else if($_POST['action'] == 'Add'){
-			$query = "INSERT INTO `game_entry`(`entryid`, `title`, `gameid`, `order`) VALUES (NULL, ?, ?, ?)";
+			$query = "INSERT INTO `character`(`idcharacter`, `Name`, `game_idgame`) VALUES (NULL,?,?)";
 			$result = $conn -> prepare($query);
-			$result -> bind_param("sii", $_POST['entry'], $_GET['gameid'], $_POST['order']);
+			//print_r($_POST);
+			$result -> bind_param("si", $_POST['character'], $_GET['gameid']);
 			$result -> execute();
 		}
 	}
@@ -66,33 +86,31 @@
 				<div class="form-group">
 					<?php
 						
-						
-						require "server/conexao.php";
-						$query = "SELECT `entryid`, `title`, `order` FROM `game_entry` WHERE `gameid` = ? ORDER BY `order`,`title`";
+						$query = "SELECT `idcharacter`, `Name` FROM `character` WHERE `game_idgame` = ? ORDER BY Name;";
 						$result = $conn -> prepare($query);
 						$result -> bind_param("i",$_GET['gameid']);
 						$result -> execute();
 						//print_r($result);
 						
 						//$game -> get_result()
-						echo '<table class="table table-hover align-middle caption-top combosuki-main-reversed text-white">';
+						echo '<table id="myTable" class="table table-hover align-middle caption-top combosuki-main-reversed text-white">';
 						echo '<tr>';
-						echo '<th>Entry</th>';
+						echo '<th>Character</th';
 						echo '</tr>';
 						foreach($result -> get_result()	as $lol){
 							
 							echo '<tr><td>';
-							echo '<form method="post" action="editentries.php?gameid='.$_GET['gameid'].'">';
-							echo '<div class="input-group"><textarea name="entry" maxlength="45" style="background-color: #474747; color:#ffffff;" class="form-control" rows="1" placeholder="Entry Title">'.$lol['title'].'</textarea>';
-							echo '<input class="form-control" type="number" name="order" placeholder="Order" value="'.$lol['order'].'" step="any">';
+							echo '<form method="post" action="editcharacters.php?gameid='.$_GET['gameid'].'">';
+							echo '<div class="input-group"><textarea name="character" maxlength="45" style="background-color: #474747; color:#ffffff;" class="form-control" rows="1" placeholder="Character Name">'.$lol['Name'].'</textarea>';
+							//echo $lol['Name'];
 							echo '
   <input name="gamePass" type="password" maxlength="16" style="background-color: #474747; color:#999999;" class="form-control" rows="1" placeholder="Game Password">
   <div class="input-group-append" id="button-addon4">
-  <input type="hidden" name="entryid" value="'.$lol['entryid'].'">
+  <input type="hidden" name="idcharacter" value="'.$lol['idcharacter'].'">
     <button type="submit" name="action" value="Update" class="btn btn-primary">Update</button>
     <button type="submit" name="action" value="Delete" class="btn btn-danger" ';
 							if(1):?>
-							onclick="return confirm('Are you sure you want to delete this entry?');"
+							onclick="return confirm('Are you sure you want to delete this character?');"
 							<?php
 							endif;
 							echo ' >Delete</button>
@@ -105,9 +123,9 @@
 						}
 						
 						echo '<tr><td>';
-							echo '<form method="post" action="editentries.php?gameid='.$_GET['gameid'].'">';
-							echo '<div class="input-group"><textarea name="entry" maxlength="45" style="background-color: #474747; color:#ffffff;" class="form-control" rows="1" placeholder="Entry Name" autofocus></textarea>';
-							echo '<input class="form-control" type="number" name="order" placeholder="Order" value="" step="any">';
+							echo '<form method="post" action="editcharacters.php?gameid='.$_GET['gameid'].'">';
+							echo '<div class="input-group"><textarea name="character" maxlength="45" style="background-color: #474747; color:#ffffff;" class="form-control" rows="1" placeholder="Character Name" autofocus></textarea>';
+							//echo $lol['Name'];
 							echo '
   <input name="gamePass" type="password" maxlength="16" style="background-color: #474747; color:#999999;" class="form-control" rows="1" placeholder="Game Password">
   <div class="input-group-append" id="button-addon4">
@@ -123,7 +141,6 @@
 						edit_controls($_GET['gameid']);
 						mysqli_close($conn);
 					?>
-					<p>Entry types are the way you categorize submissions, a few examples are: Combo, Blockstring, Okizeme...</p>
 				</div>
 			</div>
 		<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
