+++ server/connection.php
@@ -1,0 +1,10 @@
+<?php
+
+		$db_name = "test";
+		$mysql_user = "root";
+        $mysql_pass = "";
+        $server_name = "localhost:3307";
+
+        $conn = mysqli_connect($server_name, $mysql_user, $mysql_pass,$db_name);
+
+?>
