+++ server/initialize.php
@@ -1,0 +1,12 @@
+<?php
+
+	$URLDepth = isset($URLDepth) ? $URLDepth : '';
+	require_once $URLDepth."server/connection.php";
+	require_once $URLDepth."server/functions.php";
+
+
+	/*
+	$URLDepth = '../';
+	require_once "server/initialize.php";
+	*/
+?>
